+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 - touchbundle package supports creation of metric bundles, which are logical groups of metrics
+- a version of NewUntypedFunc that allows for flexible function signatures
 
 ## [v0.1.0]
 - Updated go.uber.org/fx to 1.17.1